code added in equipment module

+++ DPU_UI/src/com/dpu/controller/EquipmentEditController.java
@@ -27,7 +27,7 @@ public class EquipmentEditController extends Application implements Initializabl
 	@FXML
 	Button btnUpdateEquipment;
 	
-	int equipmentId = 0;
+	Long equipmentId = 0l;
 	
 	@FXML
 	TextField txtDescription;
